Started major rewrite, file restructuring
